+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Improved the ergonomics of handling errors from `EventProcessor`. `ProcessOutput` has a new method `result()` to get a `WeaselResult`.
 
 ## [0.9.0] - 2020-08-15
 ### Changed